+++ CHANGELOG.rst
@@ -63,7 +63,7 @@
 
 - Improvements
 
-  - Fix several typing issues with SpectrumNDCollection.
+  - Fix several typing issues with Spectrum collections.
 
 - Other changes